### 2026-01-18 #9836 @editorialbot
[PRE REVIEW]: dbparser: An R Package for Parsing and Integrating Pharmacological Databases
**Submitting author:** @MohammedFCIS (<a href="http://orcid.org/0000-0001-8907-2374">Mohammed Ali</a>)
**Repository:** https://github.com/ropensci/dbparser
**Branch with paper.md** (empty if default branch): cran
**Version:** v2.2.0
