### 2026-01-22 b919218
- Updated JOSS paper
