+++ paper.md
@@ -13,37 +13,130 @@ authors:
     orcid: 0000-0001-8907-2374
     corresponding: true
     affiliation: 1
+  - name: Ali Ezzat
+    affiliation: 1
 affiliations:
- - name: Data Analysis Engineer at Interstellar, Egypt
+ - name: Independent Researcher
    index: 1
-date: 27 December 2025
+date: 19 January 2026
 bibliography: paper.bib
 ---
 
 # Summary
 
-`dbparser` is an rOpenSci peer-reviewed R package that provides a unified framework for parsing and integrating major pharmacological and pharmacovigilance databases into standardized, analysis-ready R objects. The package currently supports three essential drug information resources: DrugBank [@wishart2018drugbank], OnSIDES [@galeano2022onsides] and TWOSIDES [@tatonetti2012data]. Each database is parsed into a consistent nested list structure called a `dvobject`, which preserves complex relational hierarchies while enabling seamless integration across databases. By providing high-performance parsing functions, chainable merge operations, and comprehensive metadata tracking, `dbparser` eliminates a significant bottleneck in computational pharmacology research and enables reproducible, large-scale drug safety analyses that would otherwise require substantial custom development effort.
+`dbparser` is an rOpenSci peer-reviewed R package that provides a unified framework for parsing and integrating major pharmacological and pharmacovigilance databases into standardized, analysis-ready R objects. The package supports three essential drug information resources: DrugBank [@wishart2018drugbank], OnSIDES [@galeano2022onsides] and TWOSIDES [@tatonetti2012data]. Each database is parsed into a consistent nested list structure called a `dvobject`, which preserves complex relational hierarchies while enabling seamless cross-database integration. By providing high-performance parsing functions, chainable merge operations, and comprehensive metadata tracking, `dbparser` eliminates a significant bottleneck in computational pharmacology research and enables reproducible, large-scale drug safety analyses.
 
 # Statement of Need
 
-Pharmacological research increasingly relies on the integration of heterogeneous data sources to understand drug mechanisms, predict adverse effects, and identify drug-drug interactions. Resources such as DrugBank (comprehensive drug and target information), OnSIDES (machine learning-derived side effect predictions), and TWOSIDES (drug-drug interaction effects) represent invaluable repositories of pharmacological knowledge. However, accessing and integrating these databases presents substantial technical challenges that impede research progress.
+Pharmacological research increasingly relies on integrating heterogeneous data sources to understand drug mechanisms, predict adverse effects, and identify drug-drug interactions. Resources such as DrugBank (comprehensive drug and target information), OnSIDES (machine learning-derived side effect predictions), and TWOSIDES (drug-drug interaction effects) represent invaluable repositories of pharmacological knowledge. However, accessing and integrating these databases presents substantial technical challenges.
 
-Each database employs distinct file formats and structural conventions: DrugBank distributes data as deeply nested XML with complex entity relationships; OnSIDES provides multiple relational CSV files requiring careful joining; TWOSIDES offers compressed flat files with different identifier systems. Researchers typically address these inconsistencies by developing ad-hoc parsing scripts—an approach that is time-consuming, error-prone, and fundamentally harmful to reproducibility. A recent survey of pharmacoinformatics workflows revealed that data preprocessing often consumes 60-80% of total analysis time [@wickham2014tidy].
[…]
