### 2026-01-22 #9836 @MohammedFCIS
Dear @arfon,

Thank you for the updated requirements and guidance. I have revised the paper to include all three required sections:

1. **Software Design** (~350 words): Details our dvobject architecture, hub-and-spoke integration model, and chainable merge operations. Includes explicit "build vs. contribute" justification comparing dbparser to Bioconductor AnnotationHub, archived drugbank R package, and Python alternatives.

2. **Research Impact Statement** (~500 words): Documents:
   - 50,000 CRAN downloads with sustained growth over 6 years
   - 10+ peer-reviewed publications in Nature, Oxford Academic, Elsevier, and MDPI journals
   - Featured in CRAN Epidemiology Task View
   - rOpenSci peer-reviewed (Issue #347, Feb 2020)
   - 6+ years of public development history (Sept 2018 - present)
   - 614 commits across 7 contributors
   - 98% test coverage and OpenSSF Best Practices passing

3. **AI Usage Disclosure**: Transparently discloses AI assistance in manuscript preparation while noting core software development (2018-2024) predates modern AI tools.

The package demonstrates the development history, quality signals, and realized research impact emphasized in JOSS's updated criteria.

Please let me know if any additional information is needed.

Best regards,
Mohammed Ali
