### 2026-01-27 #9836 @arfon
Apologies, we missed one section from the list above. In addition to the sections already mentioned, JOSS papers now also require:

Apologies again for the oversight.
