### 2026-01-28 b58ef66
- Added "state of the Field" section
