### 2026-01-28 #9836 @MohammedFCIS
Hi @arfon , may you please help me and point me to the right direction as the editorial bot is not detecting the "State of the field" section, even though:

1. The section clearly exists in the paper (line 37)
2. The heading format matches the example paper exactly: `# State of the field`
3. Other sections with the same format and subsections are being detected

Could you please check if there's an issue with the bot? The section is present and properly formatted.

Link to paper.md: [[paper.md]](https://github.com/ropensci/dbparser/blob/cran/paper.md)
Thank you
