### 2026-02-11 #9836 @arfon
@emmamendelsohn, @haozhu233 – adding you as reviewers here as you completed the rOpenSci review already. To be clear, there's no additional work required here by you, just trying to make sure you get the appropriate credit here :-)
