### 2026-02-11 #9950 @arfon
Proceeding to accept without further review as this submission was reviewed by rOpenSci here: https://github.com/ropensci/software-review/issues/347
