### 2026-02-11 #9950 @MohammedFCIS
@arfon - no I do  not have Zenodo for dbparser. Thank you
