### 2026-02-11 #9950 @arfon
OK, could you please make a new release of this software that includes the changes that have resulted from the rOpenSci review (you may have already done this). Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
