### 2026-02-11 4960a04
Fix title quotes in paper.md
