### 2026-02-12 #9950 @arfon
@shrishar – if you are interested in reviewing for JOSS, please register your interest here: https://reviewers.joss.theoj.org

While we appreciate your interest, unsolicited offers to review are not generally acceptable at JOSS.
