### 2026-02-12 #9950 @arfon
@MohammedFCIS - you have two `paper.md` files in the repo which Editorialbot cannot handle. Assuming this is not the correct one, please delete this paper https://github.com/ropensci/dbparser/blob/cran/docs/paper.md
