+++ paper.bib
@@ -29,7 +29,7 @@
   journal={medRxiv},
   year={2022},
   publisher={Cold Spring Harbor Laboratory Press},
-  doi={10.1101/2022.05.18.22275166}
+  doi={10.1101/2024.03.22.24304724}
 }
 
 @article{kuhn2016sider,
@@ -86,7 +86,7 @@
   pages={8497},
   year={2023},
   publisher={Nature Publishing Group},
-  doi={10.1038/s41598-023-35621-w}
+  doi={10.1038/s41598-023-27420-x}
 }
 
 @article{perez2021covid,
@@ -142,7 +142,7 @@
   journal={medRxiv},
   year={2024},
   publisher={Cold Spring Harbor Laboratory Press},
-  doi={10.1101/2024.01.15.24301159}
+  doi={10.1101/2024.10.17.24315553}
 }
 
 @article{rischke2023machine,
@@ -153,7 +153,7 @@
   pages={10965},
   year={2023},
   publisher={Nature Publishing Group},
-  doi={10.1038/s41598-023-37911-w}
+  doi={10.1038/s41598-023-49574-4}
 }
 
 @article{namiot2023international,
@@ -164,7 +164,7 @@
   pages={1106591},
   year={2023},
   publisher={Frontiers Media SA},
-  doi={10.3389/fphar.2023.1106591}
+  doi={10.3389/fphar.2023.1228148}
 }
 
 @article{hammoud2020multipath,
