+++ docs/paper.md
@@ -1,356 +1,0 @@
-# Summary
-
-`dbparser` is an rOpenSci peer-reviewed R package that provides a
-unified framework for parsing and integrating major pharmacological and
-pharmacovigilance databases into standardized, analysis-ready R objects.
-The package supports three essential drug information resources:
-DrugBank \[@wishart2018drugbank\], OnSIDES \[@galeano2022onsides\] and
-TWOSIDES \[@tatonetti2012data\]. Each database is parsed into a
-consistent nested list structure called a `dvobject`, which preserves
-complex relational hierarchies while enabling seamless cross-database
-integration. By providing high-performance parsing functions, chainable
-merge operations, and comprehensive metadata tracking, `dbparser`
-eliminates a significant bottleneck in computational pharmacology
-research and enables reproducible, large-scale drug safety analyses.
-
-# Statement of Need
-
-Pharmacological research increasingly relies on integrating
-heterogeneous data sources to understand drug mechanisms, predict
-adverse effects, and identify drug-drug interactions. Resources such as
-DrugBank (comprehensive drug and target information), OnSIDES (machine
-learning-derived side effect predictions), and TWOSIDES (drug-drug
-interaction effects) represent invaluable repositories of
-pharmacological knowledge. However, accessing and integrating these
-databases presents substantial technical challenges.
-
-Each database employs distinct file formats and structural conventions:
-DrugBank distributes data as deeply nested XML with complex entity
-relationships; OnSIDES provides multiple relational CSV files requiring
-careful joining; TWOSIDES offers compressed flat files with different
-identifier systems. Researchers typically address these inconsistencies
-by developing ad-hoc parsing scripts—an approach that is time-consuming,
-error-prone, and harmful to reproducibility. Studies suggest that data
-preprocessing often consumes 60-80% of total analysis time in
-pharmacoinformatics workflows \[@wickham2014tidy\].
-
-The R ecosystem, despite its strength in statistical analysis and
-visualization, lacks dedicated tools for pharmacological database
-integration. While Bioconductor \[@gentleman2004bioconductor\] provides
-excellent infrastructure for genomics data, no equivalent standardized
-framework exists for drug databases. `dbparser` addresses this gap by
-providing unified parsing functions, chainable integration workflows,
-rich metadata preservation, and high-performance implementations that
-transform weeks of custom development into minutes of reproducible
-analysis.
-
-# Software Design
-
-## Design Philosophy and Trade-offs
-
-`dbparser`’s architecture reflects three core design decisions that
-emerged from extensive experience with pharmacological data analysis
-workflows:
-
-**Unified `dvobject` Structure vs. Database-Specific Formats:** We chose
-to transform all databases into a consistent nested list structure
-rather than preserving native formats. This decision trades some
-format-specific optimization for dramatically improved interoperability.
-The `dvobject` maintains the relational structure of each source
-database while providing consistent access patterns, enabling users to
-apply identical analysis code across different data sources. Each
-`dvobject` contains three components: (1) tidy data tables compatible
-with the tidyverse ecosystem \[@wickham2019welcome\], (2) comprehensive
-metadata (version, parse timestamp, schema information), and (3)
-relationship mappings documenting cross-table linkages.
-
-**Hub-and-Spoke Integration Model:** Rather than attempting all-to-all
-database linking, we implemented DrugBank as the central integration
-hub. This reflects DrugBank’s comprehensive identifier mappings (RxCUI,
[…]
