### 2026-02-12 #9950 @MohammedFCIS
@arfon - I deleted it https://github.com/ropensci/dbparser/pull/200 (it was autogenerated by pkgdown). Thank you
