### 2026-02-12 #9950 @arfon
@MohammedFCIS – would you mind checking those invalid DOIs?
