### 2026-02-12 #9950 @MohammedFCIS
@arfon I updated the invalid DOIs and merged it in https://github.com/ropensci/dbparser/pull/202. Thank you
