### 2026-02-12 #9950 @shrishar
> [@shrishar](https://github.com/shrishar) – if you are interested in reviewing for JOSS, please register your interest here: https://reviewers.joss.theoj.org
>
> While we appreciate your interest, unsolicited offers to review are not generally acceptable at JOSS.

@arfon  I am already in the JOSS reviewer list since 2025. As of now I have reviewed 3 papers. This is not an unsolicited offer. Unfortunately lately there are so many papers are pending and none of the editors are tagging reviewers. So I decided to leave a comment in the paper I find interesting.
