### 2026-02-13 #9950 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
